python-bank-services/requirements.txt
Flask==2.3.3
pymongo==4.5.0
bson==0.5.10
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-WTF==1.1.1
Flask-Bootstrap4==4.0.2